auth/requirements.txt
anyio==3.4.0
asgi-lifespan==1.0.1
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.0
beanie==1.8.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cryptography==36.0.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.71.0
h11==0.12.0
httpcore==0.14.7
httpx==0.21.1
idna==3.3
iniconfig==1.1.1
motor==2.5.1
multidict==6.0.2
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pymongo==3.12.3
pyparsing==3.0.7
pytest==6.2.5
pytest-asyncio==0.16.0
python-jose==3.3.0
requests==2.26.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
yarl==1.7.2
gunicorn==20.1.0
